kidney_project/requirements.txt
streamlit
numpy
Pillow
scikit-learn
xgboost
joblib